requirements.txt
asgiref==3.5.2
autopep8==2.0.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-cors==0.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
networkx==2.8.8
numpy==1.23.5
packaging==22.0
Pillow==9.3.0
PyAutoGUI==0.9.53
pycodestyle==2.9.1
PyGetWindow==0.0.9
PyJWT==2.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
pytz==2022.6
pywhatkit==5.4
pywin32==305
requests==2.28.1
rotate-screen==0.1.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
wikipedia==1.4.0
